+++ NEWS.md
@@ -1,7 +1,9 @@
 # jftargets 0.0.1
 
-* Add `tar_files` as an alternative to `tarchetypes::tar_files` that does not always invalidate the upstream target.
+* Added `tar_files` as an alternative to `tarchetypes::tar_files` that does not always invalidate the upstream target.
 
-* Add `tar_png` to export a plot as a png.
+* Added `tar_png` to export a plot as a png.
 
-* Add `tar_csv` to export a dataframe as a csv.
+* Added `tar_csv` to export a dataframe as a csv.
+
+* Added `tar_xlsx` to export a dataframe as an xlsx.